DikuMUD "diku-linux.tar.gz" gamma release
This is the first release, from March 1991, as found on various MUD
source archive sites using the above name. It actually contains a Linux
port done in November 1992, and is the only copy of gamma that I could find.
What changes they made to the C files it is hard to say. Pretty much
everything in /lib and /doc has updated timestamps. Hopefully the
changes were shallow.

Any copy of gamma I could find that did NOT have the above filename, was
in fact the Linux port, only a copy that had been run once in 1994 (and
we can see what names they picked in the playerfile)